Private Documentation

Documentation for WaveformDB.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

WaveformDB

WaveformDB.WaveformDB — Module

Main module for WaveformDB.jl – is a A Julia-native package for reading, writing, processing, and plotting physiologic signal and annotation data. The core I/O functionality is based on the Waveform Database (WFDB) specifications..

This package is heavily inspired by the original WFDB Software Package, and initially aimed to replicate many of its command-line APIs. However, the projects are independent, and there is no promise of consistency between the two, beyond each package adhering to the core specifications.

WaveformDB.jl can be used to read an ECG signal e.g.:

julia> using WaveformDB

julia> rdrecord("foo.hea")

Exports:

Header

WaveformDB.adc_gain — Method

adc_gain(s::SignalSpecLine)

a getter method for the adc_gain field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

WaveformDB.adc_resolution — Method

adc_resolution(s::SignalSpecLine)

a getter method for the adc_resolution field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

WaveformDB.adc_zero — Method

adc_zero(s::SignalSpecLine)

a getter method for the adc_zero field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

WaveformDB.base_counter_value — Method

base_counter_value(h::Header)

a getter method for the basecountervalue field in the header

WaveformDB.base_date — Method

base_date(h::Header)

a getter method for the base_date field in the header

WaveformDB.base_time — Method

base_time(h::Header)

a getter method for the base_time field in the header

WaveformDB.baseline — Method

baseline(s::SignalSpecLine)

a getter method for the baseline field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

WaveformDB.block_size — Method

block_size(s::SignalSpecLine)

a getter method for the block_size field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

WaveformDB.byte_offset — Method

byte_offset(s::SignalSpecLine)

a getter method for the byte_offset field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

WaveformDB.checksum — Function

checksum(h::Header) # retrieves the checksums from a header file
checksum(h::Header,signal)  # *calculates* the checksum of a decoded symbol

WaveformDB.checksum — Method

checksum(s::SignalSpecLine)

a getter method for the checksum field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

WaveformDB.counter_frequency — Method

counter_frequency(h::Header)

a getter method for the counter_frequency field in the header

WaveformDB.description — Method

description(s::SignalSpecLine)

a getter method for the description field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

WaveformDB.filename — Method

filename(s::SignalSpecLine)

a getter method for the filename field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

WaveformDB.format — Method

format(s::SignalSpecLine)

a getter method for the format field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

WaveformDB.initial_value — Method

initial_value(s::SignalSpecLine)

a getter method for the initial_value field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

WaveformDB.nsignals — Method

nsignals(h::Header)

number of signal specs present in a header

WaveformDB.number_of_segments — Method

number_of_segments(h::Header)

a getter method for the numberofsegments field in the header

WaveformDB.number_of_signals — Method

number_of_signals(h::Header)

a getter method for the numberofsignals field in the header

WaveformDB.parentdir — Method

parentdir(h::Header)

a getter method for the parentdir field in the header

WaveformDB.record_name — Method

record_name(h::Header)

a getter method for the record_name field in the header

WaveformDB.samples_per_frame — Method

samples_per_frame(s::SignalSpecLine)

a getter method for the samplesperframe field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

WaveformDB.samples_per_signal — Method

samples_per_signal(h::Header)

a getter method for the samplespersignal field in the header

WaveformDB.sampling_frequency — Method

sampling_frequency(h::Header)

a getter method for the sampling_frequency field in the header

WaveformDB.skew — Method

skew(s::SignalSpecLine)

a getter method for the skew field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

WaveformDB.units — Method

units(s::SignalSpecLine)

a getter method for the units field in the header. can either be used on:

Header

SignalSpecLine

Vector{SignalSpecLine}

Interface

WaveformDB.rdsignal — Method

rdsignal(h::Header,physical::Bool = true)::Tuple{Int64, Matrix{Int32}}

read the samples described in the header's signal_specs field. Header - a parsed header (.hea) file

physical - optional (defaults to true) Specifies whether to return signals in physical units in the p_signal field (True), or digital units in the d_signal field (False).

Readers

Writers

utils